fix: esbuild scripts optimization
